refactor(worker): split RPC clients
